+++ CotEditor/Sources/CEAppDelegate.m
@@ -97,6 +97,7 @@
     NSDictionary *defaults = @{CEDefaultLayoutTextVerticalKey: @NO,
                                CEDefaultSplitViewVerticalKey: @NO,
                                CEDefaultShowLineNumbersKey: @YES,
+                               CEDefaultShowDocumentInspectorKey: @NO,
                                CEDefaultShowStatusBarKey: @YES,
                                CEDefaultShowStatusBarLinesKey: @YES,
                                CEDefaultShowStatusBarLengthKey: @NO,
@@ -185,7 +186,8 @@
                                CEDefaultInsertCustomTextArrayKey: @[@"<br />\n", @"", @"", @"", @"", @"", @"", @"", @"", @"", @"",
                                                                     @"", @"", @"", @"", @"", @"", @"", @"", @"", @"",
                                                                     @"", @"", @"", @"", @"", @"", @"", @"", @"", @""],
-                               CEDefaultColorCodeTypeKey:@1,
+                               CEDefaultColorCodeTypeKey: @1,
+                               CEDefaultSidebarWidthKey: @220,
                                
                                /* -------- 以下、隠し設定 -------- */
                                CEDefaultUsesTextFontForInvisiblesKey: @NO,