+++ CotEditor/Sources/constants.h
@@ -88,6 +88,7 @@ extern NSString *const CEDefaultLastVersionKey;
 extern NSString *const CEDefaultLayoutTextVerticalKey;
 extern NSString *const CEDefaultSplitViewVerticalKey;
 extern NSString *const CEDefaultShowLineNumbersKey;
+extern NSString *const CEDefaultShowDocumentInspectorKey;
 extern NSString *const CEDefaultShowStatusBarKey;
 extern NSString *const CEDefaultShowStatusBarLinesKey;
 extern NSString *const CEDefaultShowStatusBarLengthKey;
@@ -178,6 +179,7 @@ extern NSString *const CEDefaultPrintColorIndexKey;
 extern NSString *const CEDefaultInsertCustomTextArrayKey;
 extern NSString *const CEDefaultInsertCustomTextKey;
 extern NSString *const CEDefaultColorCodeTypeKey;
+extern NSString *const CEDefaultSidebarWidthKey;
 
 // hidden settings
 extern NSString *const CEDefaultUsesTextFontForInvisiblesKey;